+++ Source/Granulator.h
@@ -37,7 +37,7 @@ class Grain
 public:
    void Spawn(Granulator* owner, double time, double pos, double speedMult, double lengthInMs, double vol, double width);
    void Process(double time, ChannelBuffer* buffer, int bufferLength, float* output);
-   void DrawGrain(int idx, float x, float y, float w, float h, int bufferStart, int viewLength, int bufferLength);
+   void DrawGrain(int idx, double x, double y, double w, double h, int bufferStart, int viewLength, int bufferLength);
    void Clear() { mVol = 0; }
 
 private:
@@ -58,7 +58,7 @@ class Granulator
 public:
    Granulator();
    void ProcessFrame(double time, ChannelBuffer* buffer, int bufferLength, double offset, double speed, float* output);
-   void Draw(float x, float y, float w, float h, int bufferStart, int viewLength, int bufferLength);
+   void Draw(double x, double y, double w, double h, int bufferStart, int viewLength, int bufferLength);
    void Reset();
    void ClearGrains();
    void SetLiveMode(bool live) { mLiveMode = live; }
